+++ RELEASE_NOTES.md
@@ -46,7 +46,7 @@ predicates to all adapter files.
 * UPDATED: The LVM adapter file to require version 1.10.0 and allow the
 `setup_call_cleanup/3` predicate to be seen as a built-in predicate.
 
-* UPDATED: The Trealla Prolog adapter file to require version 1.13.51 or later.
+* UPDATED: The Trealla Prolog adapter file to require version 1.14.0 or later.
 
 * FIXED: The LVM `lvmlgt` integration script to suppress `rlwrap` warnings.